requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astor==0.8.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Cython==3.0.12
decorator==5.2.1
Django==5.2
djangorestframework==3.16.0
easyocr==1.7.2
filelock==3.18.0
fire==0.7.0
fonttools==4.57.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.10.0
paddlepaddle==3.0.0
pdf2image==1.17.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.25.5
pyparsing==3.2.3
pytesseract==0.3.13
python-bidi==0.6.6
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.2
setuptools==78.1.1
shapely==2.1.0
simsimd==6.2.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stringzilla==3.12.5
sympy==1.13.1
termcolor==3.0.1
tifffile==2025.3.30
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.9.0
